feat: end of four hour stable state
This is as far as I got in four hours. My current struggle is while
I can manually edit the nginx configuration file I can't seem to get
docker to auto deploy the conf I have written. I've tried a few
different methods which I'll outline below, all of which cause the
container to repeatedly exit.

1. triggering the Dockerfile in the nginx/ folder to copy my conf
overtop the default.conf

2. a more complex conf file which tries to setup upstream servers,
this method requires adding another service to the docker-compose
script and a new image. I was going to look into what it would take
to auto build that but decided to try simpler versions Dockerfiles
as a first attempt

I'm going to continue to work beyond the four hour mark and try a
few more avenues. The first being trying to triger and nginx restart
in my Dockerfile. If that doesn't work I'll look into building an
image via Docker-compose.