feat: feature complete version
It would help if I had read the instruction correctly :facepalm:.

So, I was trying to make it so localhost displayed the react app.
But the ask was for localhost:8080 to display it. Which just so
happened to be one of the things I tried along the way. Once I
reread the ask I had the fix read in less than 5 minutes. Took
me another two or so hours to get here but it works.